refactor: 통계객체 리팩터링
